requirements.txt
opencv-python==4.8.1.78
numpy==1.26.2
colorama==0.4.6
rplidar-roboticia==0.9.5
RPi.GPIO == 0.7.1